requirements/local.txt
-r base.txt
django-extensions
ipython
ipdb
